Add coupling infrastructure for AQM component.
Update scavenging factors input feature to allow users
to provide a default value for all tracers by setting
any of the elements of the 'fscav_aero' array in
&gfs_physics_nml to the pair "*:value" (wildcard).
Scavenging factors for individual tracers can still
be provided in the same array. The internal default
value for all AQM tracers is 0.2.